% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fns.R
\name{forestMA.PM.HK}
\alias{forestMA.PM.HK}
\title{forestMA.PM.HK}
\usage{
forestMA.PM.HK(outc, time1)
}
\arguments{
\item{outc}{secondary outcome}

\item{time1}{time1 -- either TLV or OLV}
}
\value{
Meta-analysis model with Paule-Mandel estimator and HKSJ modification
}
\description{
forestMA.PM.HK
}
