% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fns.R
\name{forestMA.PM}
\alias{forestMA.PM}
\title{forestMA.PM}
\usage{
forestMA.PM(outc, time1)
}
\arguments{
\item{time1}{}
}
\value{
Meta-analysis model with Paule-Mandel estimator
}
\description{
forestMA.PM
}
