% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fns.R
\name{forestMA}
\alias{forestMA}
\title{forestMA}
\usage{
forestMA(outc, time1)
}
\arguments{
\item{outc}{secondary outcome}

\item{time1}{time1 -- either TLV or OLV}
}
\value{
Meta-analysis model, default arguments (i.e. REML)
}
\description{
forestMA
}
