\name{val}
\alias{val}
\docType{data}
\title{
Data used in Exercise
}
\description{
a validation data of climate-growth model in Gaoligong Mountains in Hengduan Mountains.
}
\usage{data(val)}
\format{
  A data frame with 55 observations on the following 2 variables.
  \describe{
    \item{\code{actual}}{a numeric vector}
    \item{\code{recons}}{a numeric vector}
  }
}
\details{
A validation dataset of climate-growth model in Gaoligong Mountains in Hengduan Mountains.
}

\references{
Li ZS, Shi CM, Liu YB and Zhang JL. 2010. Winter drought variations based on tree-ring data in Gaoligong Mountain, Northwestern Yunnan, China, A. D. 1795-2004.
         Geochronometria(in revison)
}
\examples{
data(val)
}
\keyword{datasets}
